Initial Commit.
This first commit contains the a barebone design of the homepage. Fonts 
and color pallet will be decided on later and will be based on feedback 
from stakeholders. Placeholder Alt tags have been placed on images in 
order to meet usability specifications, and will be changed later in 
accordance to the images.